state: track substate and computed var dependencies
When __getattribute__ is used to access a Var, check whether the Var is being
requested during the processing of a ComputedVar, and if so, save the
dependency in `computed_var_dependencies`. Later, when one of the dependent
variables changes, only the computed variables that explicitly depend on it
must be re-computed.

When __getattribute__ is used to access an inherited Var from a substate, track
variable substate dependency to `substate_var_dependencies`.  Later, when one
of the dependent variables changes, mark the substate that depends on it as
dirty.

These changes work together to reduce re-calculation and improve responsiveness
of child states that depend on Vars (computed or otherwise) from their parent
states.