state: track dirty backend vars as well
Backend vars are likely to trigger ComputedVar recalculation, so ensure that
these are limited only to the dependent ComputedVars.

When determining dirty computed vars, use a recursive call that allows
ComputedVar to depend on other ComputedVar as many levels deep as needed (and
in child states) without having to recalculate everything.