config: read config overrides from environment variables
When recalculating instance field values, first look in the constants module
for a similarly named attribute and then fall back to environment variables.

This allows all config knobs to be set via environment variables, either
directly or via an existing `get_value` call in constants.

Implement `__setattr__` to save modified config values in the environment such
that subprocesses are able to access the modified value.